A big patch for tracers and composition:
* Seperating composition codes from tracer codes, several struct members in E->trace are moved to E->composition
* Add chemical buoyancy term
* Renamed thermal_buoyancy() to get_buoyancy() to reflect the fact that this function computes buoyancy according to temperature/composition/phase
* Fix and optimize the fix_*() functions
* Remove some static variables, or converte them to const
* Change the type of E->SinCos from float to double, and replace E->trace.DSinCos
* Provide default values for "required" input parameters
* Renamed tracer_restart to tracer_ic_method, also remapped the options
* Removed option for Cartesian tracer input and trace.iwrite_tracers_every
* Replaced trace.itracer_type by composition.ichemical_buoyancy
* Optionally output tracers and composition fields

+++ lib/Stokes_flow_Incomp.c
@@ -420,7 +420,7 @@ void velo_from_element(E,VV,m,el,sphere_key)
   {
 
   int a, node;
-  float sint, cost, sinf, cosf;
+  double sint, cost, sinf, cosf;
   const int dims=E->mesh.nsd;
   const int ends=enodes[E->mesh.nsd];
   const int nno=E->lmesh.nno;
